Fixed the Cmake files so they compile in Visual studio. Also did the work for HW1 and removed all warning in the code with explicit casts.

+++ HW1/HW1.cpp
@@ -62,7 +62,7 @@ void preProcess(uchar4 **inputImage, unsigned char **greyImage,
 }
 
 void postProcess(const std::string& output_file, unsigned char* data_ptr) {
-  cv::Mat output(numRows(), numCols(), CV_8UC1, (void*)data_ptr);
+  cv::Mat output((int)numRows(), (int)numCols(), (int)CV_8UC1, (void*)data_ptr);
 
   //output the image
   cv::imwrite(output_file.c_str(), output);
